feature: [Roles] Change id section placement
