+++ apps/ai-dial-admin/src/components/EntityHeaderControls/Buttons/ChangedEntityButtons.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { FC, ReactNode, useCallback, useEffect, useState } from 'react';
+import { FC, ReactNode, useEffect, useState } from 'react';
 
 import { DialNeutralButton, DialPrimaryButton } from '@epam/ai-dial-ui-kit';
 import classNames from 'classnames';
@@ -9,9 +9,6 @@ import { ButtonsI18nKey } from '@/src/constants/i18n';
 import { useIsMobileScreen } from '@/src/hooks/use-is-mobile-screen';
 import { useIsOnlyTabletScreen } from '@/src/hooks/use-is-tablet-screen';
 import { useI18n } from '@/src/locales/client';
-import { useSaveValidationContext, ValidationActionType } from '@/src/context/SaveValidationContext';
-import { useNotification } from '@/src/context/NotificationContext';
-import { showEditorErrorNotifications } from './utils';
 
 interface Props {
   disableSave?: boolean;
@@ -23,8 +20,6 @@ interface Props {
 
 const ChangedEntityButtons: FC<Props> = ({ disableSave, children, onDiscard, onSave, saveLabel }) => {
   const t = useI18n();
-  const { dispatch, jsonErrors } = useSaveValidationContext();
-  const { showNotification } = useNotification();
 
   const isTablet = useIsOnlyTabletScreen();
   const isMobile = useIsMobileScreen();
@@ -34,15 +29,6 @@ const ChangedEntityButtons: FC<Props> = ({ disableSave, children, onDiscard, onS
     setButtonsClassName(classNames((isTablet || isMobile) && 'w-1/2 flex justify-center'));
   }, [isTablet, isMobile]);
 
-  const onTryToSave = useCallback(() => {
-    if (jsonErrors?.length) {
-      const errorNotifications = showEditorErrorNotifications(jsonErrors, showNotification, t);
-      dispatch({ type: ValidationActionType.SetJsonEditorNotifications, errors: errorNotifications });
-    } else {
-      onSave?.();
-    }
-  }, [jsonErrors, showNotification, t, dispatch, onSave]);
-
   return (
     <div className="flex flex-row gap-3 p-3 lg:p-0">
       <DialNeutralButton className={buttonsClassName} label={t(ButtonsI18nKey.Discard)} onClick={onDiscard} />
@@ -50,7 +36,7 @@ const ChangedEntityButtons: FC<Props> = ({ disableSave, children, onDiscard, onS
       <DialPrimaryButton
         className={buttonsClassName}
         label={saveLabel || t(ButtonsI18nKey.Save)}
-        onClick={onTryToSave}
+        onClick={onSave}
         disabled={disableSave}
       />
     </div>
