+++ apps/ai-dial-admin/src/components/Roles/View/Properties/TabContent.tsx
@@ -1,0 +1,32 @@
+'use client';
+
+import { FC } from 'react';
+
+import EntityInfoHeader from '@/src/components/EntityHeaderControls/Info/InfoHeader';
+import { DialRole } from '@/src/models/dial/role';
+import RoleProperties from './Properties';
+
+interface Props {
+  selectedRole: DialRole;
+  names: string[];
+  isSkipRefresh: boolean;
+  onChange: (role: DialRole) => void;
+}
+
+const PropertiesTabContent: FC<Props> = ({ isSkipRefresh, names, onChange, selectedRole }) => {
+  return (
+    <div className="h-full flex flex-col w-full">
+      <EntityInfoHeader id={selectedRole.name} entity={selectedRole} />
+      <div className="flex-1 min-h-0 pt-8">
+        <RoleProperties
+          selectedRole={selectedRole}
+          names={names}
+          onChangeRole={onChange}
+          isSkipRefresh={isSkipRefresh}
+        ></RoleProperties>
+      </div>
+    </div>
+  );
+};
+
+export default PropertiesTabContent;
