+++ apps/ai-dial-admin/src/components/Roles/View/View.tsx
@@ -1,26 +1,14 @@
 'use client';
 
 import { useRouter } from 'next/navigation';
-import { FC, useCallback, useEffect, useRef, useState } from 'react';
+import { FC, useCallback, useEffect, useMemo, useRef, useState } from 'react';
 
 import { cloneDeep } from 'lodash';
 
 import { getCoreRole, removeRole, updateCoreRole, updateRole } from '@/src/app/[lang]/roles/actions';
-import AddEntitiesView from '@/src/components/AddEntitiesTab/AddEntitiesView';
-import {
-  ENTITY_COLUMNS,
-  getEntitiesForRole,
-  getRelevantKeysForRole,
-  ROLES_ENTITIES_COLUMNS,
-} from '@/src/components/AddEntitiesTab/utils';
-import EntityAudit from '@/src/components/EntityView/Audit/EntityAudit';
-import HeaderButtons from '@/src/components/EntityView/Header/HeaderButtons';
+import { JsonConfiguration } from '@/src/components/EntityHeaderControls/models';
+import SimpleEntityHeader from '@/src/components/EntityHeaderControls/SimpleHeader';
 import EntityJsonEditor from '@/src/components/EntityView/JsonEditor/JsonEditor';
-import { isSetNoLimitsHidden } from '@/src/components/EntityView/Roles/utils';
-import { getSetNoLimitsOperation } from '@/src/constants/grid-columns/actions';
-import { KEYS_COLUMNS } from '@/src/constants/grid-columns/grid-columns';
-import { EntitiesI18nKey, KeysI18nKey, TabsI18nKey } from '@/src/constants/i18n';
-import { UNLIMITED_VALUE } from '@/src/constants/role';
 import { useNotification } from '@/src/context/NotificationContext';
 import { useSaveValidationContext, ValidationActionType } from '@/src/context/SaveValidationContext';
 import { useProtectedRequest } from '@/src/hooks/use-protected-request';
@@ -29,19 +17,15 @@ import { DialApplication } from '@/src/models/dial/application';
 import { DialKey } from '@/src/models/dial/key';
 import { DialModel } from '@/src/models/dial/model';
 import { DialRole } from '@/src/models/dial/role';
-import { DialRoleLimits } from '@/src/models/dial/role-limits';
 import { DialRoute } from '@/src/models/dial/route';
 import { Toolset } from '@/src/models/dial/toolset';
-import { EntitiesGridData } from '@/src/models/entities-grid-data';
 import { ExportFormat } from '@/src/types/export';
 import { ApplicationRoute } from '@/src/types/routes';
 import { getUpdateNotificationDescription, getUpdateNotificationTitle } from '@/src/utils/entities/update-entity';
-import { getViewHeaderClassName } from '@/src/utils/entities/view';
 import { isEqualSkippingUndefined } from '@/src/utils/is-equals-entity';
 import { getErrorNotification, getSuccessNotification } from '@/src/utils/notification';
 import { EntityViewTab, getRoleTabs } from '@/src/utils/tabs/utils';
-import RoleProperties from './Properties';
-import Tabs from '@/src/components/EntityHeaderControls/Tabs/HeaderTabs';
+import TabsContent from './TabsContent';
 
 interface Props {
   originalRole: DialRole;
@@ -54,7 +38,7 @@ interface Props {
   etag: string;
 }
 
-const RolesView: FC<Props> = ({ originalRole, etag, names, models, applications, keys, toolsets, routes }) => {
+const RolesView: FC<Props> = ({ originalRole, etag, keys, ...props }) => {
   const t = useI18n();
   const router = useRouter();
   const getReqRef = useRef(useProtectedRequest());
@@ -66,13 +50,26 @@ const RolesView: FC<Props> = ({ originalRole, etag, names, models, applications,
   const [activeTab, setActiveTab] = useState(EntityViewTab.Properties);
   const [selectedRole, setSelectedRole] = useState(cloneDeep(originalRole));
   const [isChanged, setIsChanged] = useState(false);
-  const [isJsonEditorEnabled, setIsJsonEditorEnabled] = useState(false);
-  const [key, setKey] = useState(0);
+  const [isEditorEnabled, setIsEditorEnabled] = useState(false);
+
   const [isSkipRefresh, setIsSkipRefresh] = useState(true);
   const [selectedFormat, setSelectedFormat] = useState<ExportFormat>(ExportFormat.ADMIN);
   const [coreRole, setCoreRole] = useState<DialKey | null>(null);
   const entityRef = useRef(selectedRole);
 
+  const jsonConfiguration = useMemo<JsonConfiguration>(
+    () => ({
+      isEditorEnabled,
+      selectedFormat,
+      onChangeSelectedFormat: setSelectedFormat,
+      onToggleEditor: () => {
+        setSelectedFormat(ExportFormat.ADMIN);
+
+        setIsEditorEnabled((prev) => !prev);
+      },
+    }),
+    [isEditorEnabled, selectedFormat],
+  );
   useEffect(() => {
     const name = (originalRole as { name: string })?.name;
     if (!coreRole && name) {
@@ -99,18 +96,13 @@ const RolesView: FC<Props> = ({ originalRole, etag, names, models, applications,
   }, [selectedRole]);
 
   const onDiscard = useCallback(() => {
-    if (isJsonEditorEnabled) {
-      dispatch({ type: ValidationActionType.SetJsonEditor, errors: [] });
-      setIsChanged(false);
+    if (isEditorEnabled) {
       setSelectedFormat(ExportFormat.ADMIN);
-      // Due to we can't set invalid JSON as variable, we can't update entity in error state.
-      // Force JSON Editor re-render to show originalEntity on discard.
-      setKey((prevKey) => prevKey + 1);
     }
-    dispatch({ type: ValidationActionType.Reset });
+
     setIsSkipRefresh(false);
     setSelectedRole(originalRole);
-  }, [isJsonEditorEnabled, originalRole, dispatch]);
+  }, [isEditorEnabled, originalRole]);
 
   const onChangeRole = useCallback(
     (entity: DialRole, skipRefresh?: boolean) => {
@@ -118,66 +110,6 @@ const RolesView: FC<Props> = ({ originalRole, etag, names, models, applications,
       setIsSkipRefresh(!!skipRefresh);
     },
     [setSelectedRole],
-  );
-
-  const onToggleJsonEditor = useCallback(() => {
-    setSelectedFormat(ExportFormat.ADMIN);
-    setIsJsonEditorEnabled((prev) => !prev);
-  }, [setIsJsonEditorEnabled]);
-
-  const onAddEntities = useCallback(
-    (rows: EntitiesGridData[]) => {
-      const newLimits: Record<string, DialRoleLimits> = {};
-      rows.forEach((row) => {
-        const limit = row.name as string;
-        if (selectedRole.limits && selectedRole.limits[limit]) {
-          selectedRole.limits[limit].enabled = true;
-        } else {
-          newLimits[limit] = { enabled: true };
-        }
-      });
-      onChangeRole({
-        ...selectedRole,
-        limits: {
-          ...(selectedRole.limits || {}),
-          ...newLimits,
-        },
-      });
-    },
-    [onChangeRole, selectedRole],
-  );
-
-  const onRemoveEntity = useCallback(
-    (row: EntitiesGridData) => {
-      const newLimits = { ...(selectedRole.limits || {}) };
-      delete newLimits[row.name as string];
-
-      onChangeRole({
-        ...selectedRole,
-        limits: { ...newLimits },
-      });
-    },
-    [onChangeRole, selectedRole],
-  );
-
-  const onAddKeys = useCallback(
-    (rows: EntitiesGridData[]) => {
-      const newKeys = rows.map((row) => row.name as string);
-      onChangeRole({
-        ...selectedRole,
-        grantedKeys: [...(selectedRole.grantedKeys || []), ...newKeys],
-      });
-    },
-    [onChangeRole, selectedRole],
-  );
-
-  const onRemoveKey = useCallback(
-    (row: EntitiesGridData) => {
-      const newRole = cloneDeep(selectedRole);
-      newRole.grantedKeys = newRole.grantedKeys?.filter((key) => key !== row.name);
-      onChangeRole(newRole);
-    },
-    [onChangeRole, selectedRole],
   );
 
   const onSave = useCallback(() => {
@@ -203,124 +135,35 @@ const RolesView: FC<Props> = ({ originalRole, etag, names, models, applications,
     });
   }, [selectedFormat, selectedRole, originalRole.name, etag, dispatch, showNotification, t, router]);
 
-  const onChangeRoleToken = useCallback(
-    (value: number, data: DialRole, token: string) => {
-      const name = data.name as string;
-      const limits = entityRef.current.limits ?? {};
-      const updatedLimits = {
-        ...limits,
-        [name]: {
-          ...limits[name],
-          [token]: value.toString(),
-        },
-      };
-
-      const updatedEntity = {
-        ...entityRef.current,
-        limits: updatedLimits,
-      };
-
-      onChangeRole(updatedEntity, true);
-    },
-    [onChangeRole],
-  );
-
-  const onSetNoLimits = useCallback(
-    (role?: DialRole) => {
-      if (role) {
-        const limits = entityRef.current.limits ?? {};
-        const updatedLimits = {
-          ...limits,
-          [role?.name as string]: {
-            ...limits[role?.name as string],
-            day: UNLIMITED_VALUE,
-            minute: UNLIMITED_VALUE,
-            month: UNLIMITED_VALUE,
-            week: UNLIMITED_VALUE,
-          },
-        };
-
-        const updatedEntity = {
-          ...entityRef.current,
-          limits: updatedLimits,
-        };
-        onChangeRole(updatedEntity);
-      }
-    },
-    [onChangeRole],
-  );
-
   return (
     <div className="flex flex-col flex-1 min-h-0 w-full bg-layer-2 rounded p-4 pb-14 lg:pb-4 relative">
-      <div className={getViewHeaderClassName(isJsonEditorEnabled)}>
-        <Tabs
-          tabs={tabs}
-          isEditorEnabled={isJsonEditorEnabled}
-          activeTab={activeTab}
-          onChangeActiveTab={setActiveTab}
-        />
-
-        <HeaderButtons
-          view={ApplicationRoute.Roles}
-          entity={selectedRole}
-          isChanged={isChanged}
-          onDiscard={onDiscard}
-          onSave={onSave}
-          onRemove={removeRole}
-          isEditorEnabled={isJsonEditorEnabled}
-          onToggleEditor={onToggleJsonEditor}
-          selectedFormat={selectedFormat}
-          onChangeSelectedFormat={setSelectedFormat}
-        />
-      </div>
+      <SimpleEntityHeader
+        view={ApplicationRoute.Roles}
+        entity={selectedRole}
+        isChanged={isChanged}
+        onDiscard={onDiscard}
+        onSave={onSave}
+        tabs={tabs}
+        jsonConfiguration={jsonConfiguration}
+        activeTab={activeTab}
+        onChangeActiveTab={setActiveTab}
+        onRemove={removeRole}
+      />
       <div className="flex-1 overflow-auto min-h-0">
-        {isJsonEditorEnabled ? (
-          <EntityJsonEditor
-            key={key}
-            entity={selectedRole}
-            setSelectedEntity={setSelectedRole}
-            setIsChanged={setIsChanged}
-          />
+        {isEditorEnabled ? (
+          <EntityJsonEditor entity={selectedRole} setSelectedEntity={setSelectedRole} setIsChanged={setIsChanged} />
         ) : (
-          <>
-            {activeTab === EntityViewTab.Properties && (
-              <RoleProperties
-                selectedRole={selectedRole}
-                names={names}
-                onChangeRole={onChangeRole}
-                isSkipRefresh={isSkipRefresh}
-              />
-            )}
-          </>
-        )}
-        {activeTab === EntityViewTab.Entities && (
-          <AddEntitiesView
-            models={models}
-            applications={applications}
-            toolsets={toolsets}
-            routes={routes}
-            onAdd={onAddEntities}
-            onRemove={onRemoveEntity}
-            customColumns={ENTITY_COLUMNS(t)}
-            additionalColumns={ROLES_ENTITIES_COLUMNS(onChangeRoleToken)}
-            customActions={[getSetNoLimitsOperation(onSetNoLimits, isSetNoLimitsHidden)]}
-            getRelevantDataForEntity={getEntitiesForRole.bind(this, selectedRole)}
+          <TabsContent
             isSkipRefresh={isSkipRefresh}
-          />
-        )}
-        {activeTab === EntityViewTab.Keys && (
-          <AddEntitiesView
-            viewTitle={t(TabsI18nKey.Keys)}
-            customColumns={KEYS_COLUMNS(t)}
-            modalTitle={t(KeysI18nKey.AddKeys)}
-            emptyDataTitle={t(EntitiesI18nKey.NoKeys)}
+            activeTab={activeTab}
+            originalRole={entityRef.current}
+            selectedFormat={selectedFormat}
+            selectedRole={selectedRole}
             keys={keys}
-            onAdd={onAddKeys}
-            onRemove={onRemoveKey}
-            getRelevantDataForEntity={getRelevantKeysForRole.bind(this, selectedRole)}
+            onChange={onChangeRole}
+            {...props}
           />
         )}
-        {activeTab === EntityViewTab.Audit && <EntityAudit entity={selectedRole} view={ApplicationRoute.Roles} />}
       </div>
     </div>
   );
